==> Performance_benchmarks/SingleCell/Chen_dataset/2_Chen_DTU.Rmd <==
---
title: "2_Chen_DTU"
author: "Jeroen Gilis"
date: "05/11/2020"
output: html_document
---

**In order to run this script (2_Chen_DTU.Rmd), the dataset Chen_benchmark_datasets_count.Rdata (or, alternatively, Chen_benchmark_datasets_count.Rdata) is required.** This file can either be generated with the 1_Chen_prepare.Rmd script or downloaded from Zenodo.

Here we run the DTU analyses for all 6 methods on all (12) Chen benchmark datasets. Note that for DEXSeq and DRIMSeq we only run the datasets with 20 cells in each group, as these methods do not scale to large datasets. NBSplice was omitted as it does not converge on datasets with many zeroes. This code runs approximately 1.5 hours on a MacBook Pro 2018, processor; 2,3 GHz Quad-Core Intel Core i5, 16GB RAM. Most of this runtime was attributed to the DEXSeq and DRIMSeq analyses.

**If you do not want to run this script, its output can also be downloaded from Zenodo: Chen_DTU_results_count.Rdata (or, alternatively, Chen_DTU_results_scaledTPM.Rdata) **

```{r setup, include=FALSE, echo=FALSE}
require("knitr")
opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# Load DTU methods

Source file that allows running all seven DTU methods that were assessed in this paper: satuRn, DoubleExpSeq, limma diffsplice, edgeRdiffsplice, DEXSeq, DRIMseq and NBSplice.

```{r, message=FALSE, warning=FALSE}
### load libraries
library(edgeR)
library(limma)
library(DEXSeq)
library(DRIMSeq)
library(DoubleExpSeq)
library(NBSplice)
library(satuRn)
library(doMC)

source(file="./Performance_benchmarks/DTU_methods.R")
```

# Set up parallel execution

Note that the optimal parameter setting for parallelizing the DTU analyses depends on your specific computing system.

```{r}
if(TRUE) {
    nrCores <- 2

    if(nrCores != 1) {
        doParallel <- TRUE
        doProgress <- 'none'

        registerDoMC(cores = nrCores)
    } else {
        doParallel <- FALSE
        doProgress <- 'text'
    }
}
```

# Run all analyses 

Run the DTU analyses for all 6 methods on all (12) Chen benchmark datasets. Note that for DEXSeq and DRIMSeq we only run the datasets with 20 cells in each group, as these methods do not scale to large datasets. NBSplice was omitted as it does not converge on datasets with many zeroes.
This code runs approximately 1.5 hours on a MacBook Pro 2018, processor; 2,3 GHz Quad-Core Intel Core i5, 16GB RAM. Most of this runtime was attributed to the DEXSeq and DRIMSeq analyses.

```{r}
### Load benchmark data
#load(file="./Data/Chen_benchmark_datasets_count.Rdata")
load(file="./Data/Chen_benchmark_datasets_scaledTPM.Rdata") 

### Combine
ChenBenchmarkData <- c(ChenBenchmarkLenient,ChenBenchmarkStringent)

### Run DTU analyses on benchmark data

print("start satuRn")

tStart <- Sys.time()
suppressWarnings(ChenDtuBenchmark_satuRn <- plyr::llply(
    .data = ChenBenchmarkData,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- satuRn_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME
        )]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
))
difftime(Sys.time(), tStart)

print("start edgeR_diffsplice")

tStart <- Sys.time()
ChenDtuBenchmark_edgeRdiffsplice <- plyr::llply(
    .data = ChenBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- edgeR_diffsplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start DoubleExpSeq")

tStart <- Sys.time()
ChenDtuBenchmark_DoubleExpSeq <- plyr::llply(
    .data = ChenBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- DoubleExpSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )
        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start limma diffsplice")

tStart <- Sys.time()
ChenDtuBenchmark_limmaDiffsplice <- plyr::llply(
    .data = ChenBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- limma_diffsplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )
        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start DEXSeq")

# DEXSeq 
tStart <- Sys.time()
ChenDtuBenchmark_DEXSeq <- plyr::llply(
    .data = ChenBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        
        if(ncol(localData$data) > 40) {
            return(NULL)
        }
      
        ### Perform DTU analysis
        localRes <- DEXSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )
        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME
        )]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start DRIMSeq")

tStart <- Sys.time()
ChenDtuBenchmark_DRIMSeq <- plyr::llply(
    .data = ChenBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
      
        if(ncol(localData$data) > 40) {
            return(NULL)
        }
      
        ### Perform DTU analysis
        localRes <- DRIMSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )
        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

### add method name to list names for easy post-analysis
names(ChenDtuBenchmark_satuRn) <- paste0('satuRn_'    , names(ChenDtuBenchmark_satuRn))
names(ChenDtuBenchmark_limmaDiffsplice) <- paste0('limma_diffsplice_', names(ChenDtuBenchmark_limmaDiffsplice))
names(ChenDtuBenchmark_DEXSeq)          <- paste0('DEXSeq_'         , names(ChenDtuBenchmark_DEXSeq))
names(ChenDtuBenchmark_DRIMSeq)         <- paste0('DRIMSeq_'        , names(ChenDtuBenchmark_DRIMSeq))
names(ChenDtuBenchmark_edgeRdiffsplice)   <- paste0('edgeR_diffsplice_'        , names(ChenDtuBenchmark_edgeRdiffsplice))
names(ChenDtuBenchmark_DoubleExpSeq)   <- paste0('DoubleExpSeq_'        , names(ChenDtuBenchmark_DoubleExpSeq))

### Save result
save(
    ChenDtuBenchmark_satuRn,
    ChenDtuBenchmark_limmaDiffsplice,
    ChenDtuBenchmark_DEXSeq,
    ChenDtuBenchmark_DRIMSeq,
    ChenDtuBenchmark_edgeRdiffsplice,
    ChenDtuBenchmark_DoubleExpSeq,
    file="./Data/Chen_DTU_results_count.Rdata")

# save(
#     ChenDtuBenchmark_satuRn,
#     ChenDtuBenchmark_limmaDiffsplice,
#     ChenDtuBenchmark_DEXSeq,
#     ChenDtuBenchmark_DRIMSeq,
#     ChenDtuBenchmark_edgeRdiffsplice,
#     ChenDtuBenchmark_DoubleExpSeq,
#     file="./Data/Chen_DTU_results_scaledTPM.Rdata")
```
