==> Performance_benchmarks/SingleCell/Chen_dataset/3_Chen_visualize.Rmd <==
---
title: "3_Chen_visualize"
author: "Jeroen Gilis"
date: "08/12/2020"
output: html_document
---

**In order to run this script (3_Chen_visualize.Rmd), the datasets Chen_benchmark_datasets_count.Rdata and Chen_DTU_results_count.Rdata (or, alternatively, Chen_benchmark_datasets_scaledTPM.Rdata and Chen_DTU_results_scaledTPM.Rdata) are required.** These files can either be generated with the 1_Chen_prepare.Rmd and 2_Chen_DTU.Rmd scripts, respectively, or downloaded from Zenodo and put in the data folder of this GitHub repository.

The results of this script (FDR-TPR curves) are already available from our GitHub, under the directory ./Results/Chen_benchmark/.

```{r setup, include=FALSE, echo=FALSE}
require("knitr")
opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# Load libraries

```{r,message=FALSE,warning=FALSE}
library(ggplot2)
library(iCOBRA)
```

# Load truth file

```{r}
load(file="./Data/Chen_benchmark_datasets_count.Rdata")
#load(file="./Data/Chen_benchmark_datasets_scaledTPM.Rdata")
metaInfo <- c(ChenBenchmarkLenient,ChenBenchmarkStringent)
rm(ChenBenchmarkLenient,ChenBenchmarkStringent)
```

# Load DTU performance results

```{r}
load(file="./Data/Chen_DTU_results_count.Rdata")
#load(file="./Data/Chen_DTU_results_scaledTPM.Rdata")

ChenBenchmark <- c(
    ChenDtuBenchmark_satuRn,
    ChenDtuBenchmark_limmaDiffsplice,
    ChenDtuBenchmark_DEXSeq,
    ChenDtuBenchmark_DRIMSeq,
    ChenDtuBenchmark_edgeRdiffsplice,
    ChenDtuBenchmark_DoubleExpSeq
)

### Remove empty entries (due to not tested - aka to many samples for reasonable runtime)
ChenBenchmark  <- ChenBenchmark[which(
    sapply(ChenBenchmark, function(x) ! is.null(x$dtuAnalysis)))]

rm(ChenDtuBenchmark_satuRn,
    ChenDtuBenchmark_limmaDiffsplice,
    ChenDtuBenchmark_DEXSeq,
    ChenDtuBenchmark_DRIMSeq,
    ChenDtuBenchmark_edgeRdiffsplice,
    ChenDtuBenchmark_DoubleExpSeq)
invisible(gc())
```

# Helper function 
To get pvalues for the different methods on the different datasets

```{r}
get_pvalues <- function(dataset,nrep,nmethods) {
  
  get_all_txs <- function(dataset,nrep){
  
    all_txs <- c()
    for (i in 1:nrep) {
      current_txs <- as.character(dataset[[i]]$dtuAnalysis$TXNAME)
      all_txs <- c(all_txs, current_txs)
      all_txs <- unique(all_txs)
    }
    return(all_txs)
  }

  all_txs <- get_all_txs(dataset,nrep)
  pvalues <- matrix(data=NA, nrow=length(all_txs), ncol=nmethods*nrep)
  rownames(pvalues) <- all_txs
  
  for (i in 1:(nmethods*nrep)) {
    pvalues_new <- data.frame(dataset[[i]]$dtuAnalysis$p_value)
    rownames(pvalues_new) <- dataset[[i]]$dtuAnalysis$TXNAME
    colnames(pvalues_new) <- names(dataset)[i]

    pvalues[,i] <- pvalues_new[match(rownames(pvalues),rownames(pvalues_new)),1]
  }
  
  pvalues <- apply(pvalues, 2, as.numeric)
  rownames(pvalues) <- all_txs
  pvalues <- as.data.frame(pvalues)
  pvalues <- pvalues[order(rownames(pvalues)),]
  
  if(nmethods == 6){
    colnames(pvalues) <- c("satuRn", "limma_diffsplice", "DEXSeq", "DRIMSeq","edgeR_diffsplice", "DoubleExpSeq")
  } else {
    colnames(pvalues) <- c("satuRn", "limma_diffsplice","edgeR_diffsplice", "DoubleExpSeq")
  }
  return(pvalues)
}
```

# Visualization function

```{r}
visualize_datasets <- function(dataset,metaInfo,nFilters,nSampleSizes,nRep,nmethods,selection,selection2) {

    count <- 0
    nPlots <- nFilters * nSampleSizes
    print(paste("This function will generate",  nPlots, "plots"))
  
    for (j in 1:nPlots) {
        print(j)
        Chen_current_outer <- dataset[grepl(selection[j], names(dataset))]
        metaInfo_current_outer <- metaInfo[grepl(selection2[j], names(metaInfo))]

        resList <- c()
        count <- count+1

        for (k in 1:length(nRep)) {
            Chen_current_inner <- Chen_current_outer[grepl(nRep[k], names(Chen_current_outer))]

            metaInfo_current_inner <- metaInfo_current_outer[grepl(nRep[k], names(metaInfo_current_outer))]
    
            pvalues <- get_pvalues(Chen_current_inner,nrep=1,nmethods)

            truth_file <- Chen_current_inner[[1]]$dtuAnalysis
    
            if(count <= nSampleSizes){
                txSwapped <- metaInfo_current_inner[1]          
                } else {
                txSwapped <- metaInfo_current_inner[2]
            }
    
            truth_file <- truth_file[,c("TXNAME","p_value")]
    
            txSwapped[[1]]$metaInfo <- txSwapped[[1]]$metaInfo[match(truth_file$TXNAME,txSwapped[[1]]$metaInfo$TXNAME),]
    
            truth_file$gene_modified <- as.numeric(txSwapped[[1]]$metaInfo$txSwapped)
            truth_file <- truth_file[order(truth_file$TXNAME),]
            rownames(truth_file) <- truth_file$TXNAME

            resList[[k]] <- list(pvalues,truth_file)
        }
        
        rownames(resList[[1]][[1]]) <- paste0(rownames(resList[[1]][[1]]), "_1")
        rownames(resList[[2]][[1]]) <- paste0(rownames(resList[[2]][[1]]), "_2")
        rownames(resList[[3]][[1]]) <- paste0(rownames(resList[[3]][[1]]), "_3")
        
        pvalues_full <- rbind(resList[[1]][[1]],resList[[2]][[1]],resList[[3]][[1]])
        
        rownames(resList[[1]][[2]]) <- paste0(rownames(resList[[1]][[2]]), "_1")
        rownames(resList[[2]][[2]]) <- paste0(rownames(resList[[2]][[2]]), "_2")
        rownames(resList[[3]][[2]]) <- paste0(rownames(resList[[3]][[2]]), "_3")
        
        truth_file_full <- rbind(resList[[1]][[2]],resList[[2]][[2]],resList[[3]][[2]])
        
        cobra <- COBRAData(pval = pvalues_full, truth = truth_file_full)
        cobra <- calculate_adjp(cobra)
        cobra1perf <- calculate_performance(cobra, binary_truth = "gene_modified", cont_truth = "none", splv = "none", aspects = c("fdrtpr", "fdrtprcurve", "overlap"))

        cobraplot <- prepare_data_for_plot(cobra1perf, colorscheme = "Dark2", facetted = TRUE)
        
        if (nmethods == 6){
            new_col <- c(rep(c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "black", "#999999"),4),rep("white",14))
            names(new_col) <- names(cobraplot@plotcolors)
            cobraplot@plotcolors <- new_col
    
            plot <- plot_fdrtprcurve(cobraplot, xaxisrange = c(0, 0.4), yaxisrange = c(0,1))
    
            titles <- c("20 versus 20 - edgeR filter - count","20 versus 20 - DRIMSeq filter - count")
            #titles <- c("20 versus 20 - edgeR filter - scaledTPM","20 versus 20 - DRIMSeq filter - scaledTPM")
            current_title <- titles[j]
        } else {
            new_col <- c(rep(c("#56B4E9", "#F0E442", "#0072B2", "black", "#999999"),4),rep("white",10))
            names(new_col) <- names(cobraplot@plotcolors)
            cobraplot@plotcolors <- new_col
    
            plot <- plot_fdrtprcurve(cobraplot, xaxisrange = c(0, 0.4), yaxisrange = c(0,1))
    
            titles <- c("50 versus 50 - edgeR filter - count","50 versus 50 - DRIMSeq filter - count")
            #titles <- c("50 versus 50 - edgeR filter - scaledTPM","50 versus 50 - DRIMSeq filter - scaledTPM")
            current_title <- titles[j]
        }

        plot <- plot + 
            ggtitle(current_title) +
            theme(strip.background = element_blank(),
            strip.text.x = element_blank(),
            plot.title = element_text(size=10),
            axis.text.x = element_text(size = 10),
            axis.text.y = element_text(size = 10),
            axis.title.x = element_text(size = 10),
            axis.title.y = element_text(size = 10))
        

      png(paste("./Results/Chen_benchmark/FDRTPR_", selection2[j], "_", sub("^[^f]*", "", selection[j]), "_count.png", sep = ""),
            width     = 5.5,
            height    = 4.5,
            units     = "in",
            res       = 200,
            pointsize = 4) # start export
      print(plot)
      dev.off()

      # png(paste("./Results/Chen_benchmark/FDRTPR_", selection2[j], "_", sub("^[^f]*", "", selection[j]), "_scaledTPM.png", sep = ""),
      #       width     = 5.5,
      #       height    = 4.5,
      #       units     = "in",
      #       res       = 200,
      #       pointsize = 4) # start export
      # print(plot)
      # dev.off()
  }
}
```

# Generate FDR-TPR curves

```{r}
selection <- c("20_rep_._filterLenient","20_rep_._filterStringent")
selection2 <- c("20_rep","20_rep")

visualize_datasets(ChenBenchmark,
                   metaInfo,
                   nFilters=2,
                   nSampleSizes=1,
                   nRep=c("rep_1","rep_2","rep_3"),
                   nmethods=6,
                   selection = selection,
                   selection2=selection2)


selection <- c("_50_rep_._filterLenient","_50_rep_._filterStringent")
selection2 <- c("50_rep","50_rep")

visualize_datasets(ChenBenchmark,
                   metaInfo,
                   nFilters=2,
                   nSampleSizes=1,
                   nRep=c("rep_1","rep_2","rep_3"),
                   nmethods=4,
                   selection = selection,
                   selection2=selection2)
```
